+++ nest/main.ts
@@ -3,7 +3,7 @@ import { ValidationPipe } from '@nestjs/common';
 import cookieParser from 'cookie-parser';
 import morgan from 'morgan';
 import helmet from 'helmet';
-import AppModule from './modules/app.module';
+import AppModule from './src/modules/app.module';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
