--- nest/modules/user.module.ts
@@ -1,17 +1,0 @@
-import { Module } from '@nestjs/common';
-import UserRepository from '@models/repositories/user.repository';
-import UserService from '@services/user.service';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import UserController from '@controllers/user.controller';
-import UserJobRepository from '@models/repositories/ user-job.reposity';
-import UserFieldRepository from '@models/repositories/user-field.repository';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([UserRepository, UserJobRepository, UserFieldRepository])],
-  controllers: [UserController],
-  providers: [UserService],
-  exports: [UserService],
-})
-class UserModule {}
-
-export default UserModule;
