+++ nest/src/controllers/chat.controller.ts
@@ -1,5 +1,5 @@
 import { BadRequestException, Controller, Get, Param, ParseIntPipe } from '@nestjs/common';
-import ChatService from '@services/chat.service';
+import ChatService from '../services/chat.service';
 
 @Controller('chat')
 class ChatController {
