+++ nest/src/models/entities/helper/abstract.ts
