+++ nest/src/models/entities/room.entity.ts
