+++ nest/src/models/entities/user.entity.ts
