+++ nest/src/models/entities/users-job.entity.ts
