+++ nest/src/models/repositories/user-field.repository.ts
@@ -1,5 +1,5 @@
 import { EntityRepository, Repository } from 'typeorm';
-import UserFieldEntity from '@models/entities/user-field.entity';
+import UserFieldEntity from '../entities/user-field.entity';
 
 @EntityRepository(UserFieldEntity)
 export default class UserFieldRepository extends Repository<UserFieldEntity> {}
