+++ nest/src/modules/app.module.ts
@@ -3,7 +3,7 @@ import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import Joi from 'joi';
 import { join } from 'path';
-import ChatModule from '@modules/chat.module';
+import ChatModule from './chat.module';
 import AuthModule from './auth.module';
 import MailModule from './mail.module';
