+++ nest/src/modules/chat.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import RoomRepository from '@models/repositories/room.repository';
-import ChatController from '@controllers/chat.controller';
-import ChatGateway from '@controllers/chat.gateway';
-import ChatService from '@services/chat.service';
-import AuthModule from '@modules/auth.module';
-import UserModule from '@modules/user.module';
+import RoomRepository from '../models/repositories/room.repository';
+import ChatController from '../controllers/chat.controller';
+import ChatGateway from '../controllers/chat.gateway';
+import ChatService from '../services/chat.service';
+import AuthModule from './auth.module';
+import UserModule from './user.module';
 
 @Module({
   imports: [TypeOrmModule.forFeature([RoomRepository]), AuthModule, UserModule],
