+++ nest/src/services/passport/jwt.refresh.strategy.ts
@@ -2,9 +2,9 @@ import { Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Request } from 'express';
 import { ExtractJwt, Strategy } from 'passport-jwt';
-import AuthService from '@services/auth.service';
 import { ConfigService } from '@nestjs/config';
 import { IJwtPayload } from '@typing/auth';
+import AuthService from '../auth.service';
 
 /**
  *
