+++ nest/src/services/user.service.ts
@@ -1,9 +1,9 @@
 import { Injectable, InternalServerErrorException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import UserFieldEntity from '@models/entities/user-field.entity';
-import UserJobEntity from '@models/entities/users-job.entity';
-import UserFieldRepository from '@models/repositories/user-field.repository';
-import UserJobRepository from '@models/repositories/ user-job.reposity';
+import UserFieldEntity from '../models/entities/user-field.entity';
+import UserJobEntity from '../models/entities/users-job.entity';
+import UserFieldRepository from '../models/repositories/user-field.repository';
+import UserJobRepository from '../models/repositories/ user-job.reposity';
 import UserRepository from '../models/repositories/user.repository';
 
 @Injectable()
